Added support for multiple chats. Disabled Cleanup function for now.
