+++ src/Moneo.Functions/TaskManager/MoneoTaskCreateModel.cs
@@ -1,0 +1,6 @@
+﻿using Moneo.Models;
+
+namespace Moneo.Functions
+{
+    public sealed record MoneoTaskCreateModel(long ChatId, MoneoTaskDto TaskDto);
+}
